Experiment with undertow as embedded servlet container
The main deployment scenario is still as war in a servlet container.
First steps are taken towards deployment as fat jar with undertow as
embedded servlet engine. With fat-jar, I hope to make the project more
accessible. Undertow is not working yet, because not all resources can
be found on the classpath.